requirements.txt
python-can==3.3.4
python-benedict==0.24.0
rich==10.3.0
cantools
